[refactoring] Cars, RacingGameRecord
