+++ src/main/java/step3/game/domain/Cars.java
@@ -14,7 +14,7 @@ public class Cars {
         this(cars.cars);
     }
 
-    private Cars(List<Car> cars) {
+    public Cars(List<Car> cars) {
         this.cars = cars;
     }
 
@@ -39,6 +39,20 @@ public class Cars {
         return new Cars(newCars);
     }
 
+    public Cars winners() {
+        final int winnerPosition = getWinnerPosition();
+        return new Cars(cars.stream()
+                            .filter(car -> car.isWinner(winnerPosition))
+                            .collect(Collectors.toList()));
+    }
+
+    private int getWinnerPosition() {
+        return cars.stream()
+                   .mapToInt(Car::getPosition)
+                   .max()
+                   .orElse(0);
+    }
+
     public List<Car> getCars() {
         return cars.stream()
                    .map(Car::new)
@@ -49,6 +63,10 @@ public class Cars {
         if (numberOfCar == 0) {
             throw new IllegalArgumentException("경주 자동차는 1대 이상 추가해야 합니다.");
         }
+    }
+
+    public int getNumberOfWinners() {
+        return cars.size();
     }
 
 }
