+++ src/main/java/step3/game/dto/RacingGameResultModel.java
@@ -3,13 +3,12 @@ package step3.game.dto;
 import java.util.List;
 
 import step3.game.domain.Cars;
-import step3.game.domain.Winners;
 
 public class RacingGameResultModel {
     private final List<Cars> cars;
-    private final Winners winners;
+    private final Cars winners;
 
-    public RacingGameResultModel(List<Cars> cars, Winners winners) {
+    public RacingGameResultModel(List<Cars> cars, Cars winners) {
         this.cars = cars;
         this.winners = winners;
     }
@@ -18,7 +17,7 @@ public class RacingGameResultModel {
         return cars;
     }
 
-    public Winners getWinners() {
+    public Cars getWinners() {
         return winners;
     }
 }
